% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countReads.R
\name{countReads}
\alias{countReads}
\title{Count and summarize long-reads RNA into a table}
\usage{
countReads(alnsFile, referencesFile)
}
\arguments{
\item{alnsFile}{The file that store the long-read RNA alignments generated by Minimap2.}

\item{referencesFile}{A file that store reference coordinates of human genes.}
}
\value{
A list of two data frames. Fisrt data frame contains new exon coordinates
created and second data frame contains the exon combinations of each read.
}
\description{
A function that ouputs two tsv files containing counts of exon combinations
}
\examples{
alignments <- system.file("extdata", "mlx_reads.sorted.bam",
package = "LSplicing")
referenceCoord <- system.file("extdata", "example_refCoord.gff3",
package = "LSplicing")

\dontrun{
countReads(alns = alignments, refCoord = referenceCoord)
}
}
