% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotReads.R
\name{plotReads}
\alias{plotReads}
\title{plotReads.R}
\usage{
plotReads(coor, reads, gene, alnsFile)
}
\arguments{
\item{coor}{A data frame that contains exon coordinates of different genes.}

\item{reads}{A data frame containing rows of reads. If a value = 1,
then that row of read is mapped to that column of exon. If a value = 0, then that row
of read is not mapped to that column of exon.}

\item{gene}{A string of characters representing Ensembl gene ID of the gene of interest.}

\item{alnsFile}{The file that store the long-read RNA alignments generated by Minimap2.}
}
\value{
A ggplot graph that represent the long-read RNAs exon combinations
}
\description{
A function that plots results exon combination of target gene (plots results of countReads function).
}
\examples{
alignments <- system.file("extdata", "mlx_reads.sorted.bam",
package = "LSplicing")
referenceCoord <- system.file("extdata", "example_refCoord.gff3",
package = "LSplicing")

\dontrun{
results <- countReads(alignments, referenceCoord)
coor <- results[[1]]
r <- results[[2]]
plotReads(coor = coor,
          reads = r,
          gene = "ENSG00000108788.11",
          alns = alignments)
}
}
