% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCoverage.R
\name{readCoverage}
\alias{readCoverage}
\title{Calculates coverage of long-reads RNA}
\usage{
readCoverage(coor, alns)
}
\arguments{
\item{coor}{A data frame containing exon coordinates of target gene.}

\item{alns}{A GAlignment object of long-read RNA alignments generated by Minimap2.}
}
\value{
A data frame that stores the coverage of the reads
}
\description{
A function that calculates the coverage of long-read RNA
}
\examples{
alignments <- system.file("extdata", "mlx_reads.sorted.bam",
package = "LSplicing")
referenceCoord <- system.file("extdata", "example_refCoord.gff3",
package = "LSplicing")

\dontrun{
results <- countReads(alignments, referenceCoord)
coor <- results[[1]]
coor <- coor[coor$gene_id == "ENSG00000108788.11", ]
readCoverage(coor = coor,
             alns = alignments)
}
}
