% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLSplicingApp.R
\name{runLSplicingApp}
\alias{runLSplicingApp}
\title{\code{runLSplicingApp} run the shiny app \code{./inst/shiny-scripts/lsplicingApp/}.}
\usage{
runLSplicingApp()
}
\value{
NULL (invisible) Invoked for its side-effect of launching a shiny app.
}
\description{
\code{runLSplicingApp} run the shiny app \code{./inst/shiny-scripts/lsplicingApp/}.
}
\examples{
# No runnable example applies - but BiocCheck() requires one. So ...
NULL
\dontrun{
runLSplicingApp()
}
}
